requirements.txt
openai
openai-whisper
setuptools-rust
pyaudio
python-dotenv
elevenlabslib
replicate
opencv-python
ffpyplayer
groq
